pyproject.toml
[project]
name = "advent-of-code-2024"
version = "0.1.0"
description = ""
authors = [{ name = "Shen Li", email = "shen@shenli.ca" }]
requires-python = ">=3.12,<3.13"
dependencies = [
    "numpy>=2.1.3,<3",
    "networkx>=3.4.2,<4",
    "matplotlib>=3.9.3,<4",
    "patina>=0.5.0,<0.6",
    "pytest>=8.3.3,<9",
    "ipython>=8.30.0,<9",
    "python-dateutil>=2.9.0.post0,<3",
    "python-dotenv>=1.0.1,<2",
    "requests>=2.32.3,<3",
    "pydash>=8.0.4,<9",
    "more-itertools>=10.5.0,<11",
]

[dependency-groups]
dev = [
    "black>=24.10.0,<25",
    "isort>=5.13.2,<6",
]

[tool.uv]
package = false

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.isort]
profile = "black"
